Implemented changes necessary for a Heroku deploy.
Now by modifying a couple of variables in settings.py, we can deploy to:
- Localhost manually.
- Localhost with docker.
- Localhost with heroku local.
- Remotely to Heroku.

I decided to commit React's build folder to version control, because
that's just the easiest way to deploy it to Heroku. I don't like
it, but c'est la vie.

Note that now the requirements.txt file is used exclusively by Heroku.
Do NOT rely on it for a localhost manual deploy because it will fail,
given that the PostgreSQL Python adapter (psycopg2) is not listed.