Fixed frontend bug and minor UI improvement.
Bug fixed: Before, when a task was added and the backend was down, the
task adder widget was reset, resulting in user info loss.
Now, the task adder widget is only reset if the task was successfully added in the backend.

Minor UI improvement: The logout button now has a margin bottom.